+++ pages/profile/index.vue
@@ -7,6 +7,7 @@
           <div class="md-layout">
             <div class="md-layout-item md-size-50 mx-auto">
               <div class="profile">
+                <!--IMAGE-->
                 <div class="avatar">
                   <img
                     :src="img"
@@ -14,11 +15,17 @@
                     class="img-raised rounded-circle img-fluid"
                   >
                 </div>
+                <!--IMAGE-->
+                <!--Profile name + links-->
                 <div class="name">
+                  <!--Name + job-->
                   <h3 class="title">
                     Robbie Verdurme
                   </h3>
                   <h6>Developer</h6>
+                  <!--/Name + job-->
+
+                  <!--Facebook-->
                   <md-button
                     href="https://www.facebook.com/robbie.verdurme"
                     class="md-just-icon md-simple md-facebook"
@@ -28,6 +35,9 @@
                       {{ $t('profile.facebook') }}
                     </md-tooltip>
                   </md-button>
+                  <!--/Facebook-->
+
+                  <!--Github-->
                   <md-button
                     href="https://github.com/RobbieVerdurme"
                     class="md-just-icon md-simple md-github"
@@ -37,6 +47,9 @@
                       {{ $t('profile.github') }}
                     </md-tooltip>
                   </md-button>
+                  <!--/Github-->
+
+                  <!--Linked in-->
                   <md-button
                     href="https://www.linkedin.com/in/robbie-verdurme/"
                     class="md-just-icon md-simple md-linkedin"
@@ -46,15 +59,21 @@
                       {{ $t('profile.linkedin') }}
                     </md-tooltip>
                   </md-button>
+                  <!--/Linked in-->
                 </div>
+                <!--/Profile name + links-->
               </div>
             </div>
           </div>
+          <!--Profile info-->
           <div class="description text-center">
             <p>
               {{ $t('profile.intro') }}
             </p>
           </div>
+          <!--/Profile info-->
+
+          <!--TABS-->
           <div class="profile-tabs">
             <tabs
               :tab-name="[$t('profile.lifeTab'), $t('profile.projectsTab'), $t('profile.resumeTab')]"
@@ -72,7 +91,10 @@
 
               <!--Projects-->
               <template slot="tab-pane-2">
-                <projects />
+                <projects :items="projects" />
+                <md-button :to="{name:'projects'}" class="md-success">
+                  {{ $t('profile.allProjects') }}
+                </md-button>
               </template>
               <!--/Projects-->
 
@@ -83,6 +105,7 @@
               <!--/Resume-->
             </tabs>
           </div>
+          <!--/TABS-->
         </div>
       </div>
     </div>
@@ -92,16 +115,20 @@
 <script>
 export default {
   components: {
-    Parallax: () => import('~/components/molecules/parallax'),
-    Tabs: () => import('~/components/molecules/tabs'),
+    parallax: () => import('~/components/molecules/parallax'),
+    tabs: () => import('~/components/molecules/tabs'),
     profileInfo: () => import('~/components/molecules/profileinfo'),
     projects: () => import('~/components/organisms/projects'),
     resume: () => import('~/components/molecules/resume.vue')
   },
   data () {
     return {
+      // images
       header: 'img/background.jpg',
-      img: 'img/faces/ProfilePic.jpg'
+      img: 'img/faces/ProfilePic.jpg',
+
+      // project data
+      projects: this.$t('profile.projects').slice(0, 2)
     }
   },
   computed: {
@@ -112,6 +139,14 @@ export default {
       return {
         backgroundImage: `url(${this.header})`
       }
+    }
+  },
+  watch: {
+    /**
+     * update projects with correct lang
+     */
+    '$i18n.locale' () {
+      this.projects = this.$t('profile.projects').slice(0, 2)
     }
   }
 }