+++ store/state.js
@@ -1,5 +1,3 @@
 export default () => ({
-  locales: ['en', 'fr', 'nl'],
-  locale: 'en',
   navbar: false
 })